Added support for loginout in the backend
